pyproject.toml
[project]
name = "eosin"
version = "0.1.0"
description = "Standalone GLM OCR bank parser backend for Modal deployment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate",
    "beautifulsoup4",
    "blinker",
    "camelot-py[cv]",
    "eliot",
    "fastapi",
    "glmocr[selfhosted,server]",
    "img2table",
    "liteparse",
    "modal>=1.4.2",
    "numpy",
    "opencv-python-headless",
    "pandas",
    "pillow",
    "portalocker",
    "pydantic",
    "pymupdf",
    "python-dotenv",
    "python-multipart",
    "pyyaml",
    "requests",
    "sentencepiece",
    "torch",
    "huggingface-hub==1.4.1",
    "transformers==5.3.0",
    "tqdm",
    "uvicorn",
]
